* ext/dl/cptr.c (rb_dlptr_new2): remove unused variable.
	* ext/iconv/iconv.c (get_iconv_opt_i): ditto.

	* ext/syck/token.c (Plain): fix types on LP64.


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@34400 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/syck/token.c
@@ -1489,8 +1489,8 @@ yy109:
 
 Plain:
     {
-        int qidx = 0;
-        int qcapa = 100;
+        ptrdiff_t qidx = 0;
+        ptrdiff_t qcapa = 100;
         char *qstr = S_ALLOC_N( char, qcapa );
         SyckLevel *plvl;
         int parentIndent;
